fix locations file structure
